@@ -23,13 +23,14 @@
     "extends": "react-app"
   },
   "dependencies": {
+    "@babel/core": "^7.17.0",
+    "@babel/plugin-transform-runtime": "^7.17.0",
+    "@babel/preset-env": "^7.16.11",
     "@luigi-project/client": "^1.20.1",
     "@sap-theming/theming-base-content": "^11.1.28",
-    "@sentry/react": "^6.12.0",
     "@sentry/tracing": "^6.12.0",
     "ajv": "^6.10.2",
     "classnames": "^2.2.6",
-    "constate": "^1.2.0",
     "copy-to-clipboard": "^3.2.0",
     "core-js": "^3.3.6",
     "file-saver": "^2.0.2",
@@ -46,21 +47,24 @@
     "monaco-editor": "^0.23.0",
     "pluralize": "^8.0.0",
     "prop-types": "^15.7.2",
-    "react": "16.12.0",
-    "react-dom": "16.12.0",
-    "react-i18next": "^11.11.3",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-i18next": "^11.15.3",
+    "react-router-dom": "^6.2.1",
+    "regenerator-runtime": "^0.13.9",
     "rfc6902": "^4.0.1",
     "semver": "^7.3.5"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^8.2.3",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
     "cli-color": "^1.4.0",
     "concurrently": "^5.0.0",
     "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
     "enzyme-to-json": "^3.4.0",
     "fancy-log": "^1.3.3",
     "gulp": "^4.0.1",
@@ -75,7 +79,6 @@
     "node-sass": "^7.0.1",
     "prettier": "^1.18.2",
     "react-scripts": "5.0.0",
-    "react-test-renderer": "16.9.0",
     "sass-loader": "^12.4.0",
     "waait": "^1.0.5",
     "webpack-cli": "^4.9.2",
