@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes, useParams } from 'react-router-dom';
 
 import ServiceClassList from '../ServiceClassList/ServiceClassList';
 import {
@@ -11,7 +11,6 @@ import ServiceInstancesList from '../ServiceInstanceList/ServiceInstanceList';
 import ServiceInstancesDetails from '../ServiceInstanceDetails/ServiceInstanceDetails';
 import {
   NotificationProvider,
-  withTitle,
   MainFrameRedirection,
   useMicrofrontendContext,
 } from 'react-shared';
@@ -26,49 +25,39 @@ const App = () => {
   }, [language, i18n]);
   return (
     <NotificationProvider>
-      <Switch>
-        <Route path="/preload" component={() => null} />
-
-        <Route
-          exact
-          path="/catalog"
-          render={withTitle(CATALOG_TITLE, ServiceClassList)}
-        />
+      <Routes>
+        <Route exact path="/catalog" element={<ServiceClassList />} />
 
         <Route
           exact
           path="/catalog/ServiceClass/:name"
-          render={withTitle(CATALOG_TITLE, RoutedServiceClassDetails)}
+          element={<RoutedServiceClassDetails />}
         />
 
         <Route
           exact
           path="/catalog/ClusterServiceClass/:name"
-          render={withTitle(CATALOG_TITLE, RoutedClusterServiceClassDetails)}
+          element={<RoutedClusterServiceClassDetails />}
         />
 
-        <Route
-          exact
-          path="/instances"
-          render={withTitle(INSTANCES_TITLE, ServiceInstancesList)}
-        />
+        <Route exact path="/instances" element={<ServiceInstancesList />} />
         <Route
           exact
           path="/instances/details/:name"
-          render={withTitle(INSTANCES_TITLE, ServiceInstancesDetails)}
+          element={<ServiceInstancesDetails />}
         />
-        <Route exact path="" component={MainFrameRedirection} />
-      </Switch>
+        <Route exact path="" element={<MainFrameRedirection />} />
+      </Routes>
     </NotificationProvider>
   );
 };
 
-const RoutedServiceClassDetails = ({ match }) => (
-  <ServiceClassDetailsContainer name={match.params.name} />
-);
-
-const RoutedClusterServiceClassDetails = ({ match }) => (
-  <ClusterServiceClassDetailsContainer name={match.params.name} />
-);
-
+const RoutedServiceClassDetails = () => {
+  const { name } = useParams();
+  return <ServiceClassDetailsContainer name={name} />;
+};
+const RoutedClusterServiceClassDetails = () => {
+  const { name } = useParams();
+  return <ClusterServiceClassDetailsContainer name={name} />;
+};
 export default App;
